Create DOM build without bundled Yoga
